Handling startup scripts
