+++ pytest.ini
@@ -5,6 +5,7 @@ flake8-max-line-length = 160
 flake8-ignore =
     # E402 module level import not at top of file
     setup.py D200 D400
+    start.py D102 D200 D400 D401 E402 E501
     medusa/bs4_parser.py D100 D101 D102 D105
     medusa/classes.py D100 D101 D102 D105 N802 N803
     medusa/common.py D101 D102 D105 D200 D202 D204 D205 D400 D401 E501 N802 N812
@@ -129,7 +130,6 @@ flake8-ignore =
     medusa/providers/torrent/xml/__init__.py D104
     medusa/providers/torrent/xml/kat.py D100 D102 D202 D204 D400 E501
     medusa/providers/torrent/xml/torrentz2.py D100 D102 D202 D204 D400
-    SickBeard.py D102 D200 D400 D401 E402 E501
     medusa/rssfeeds.py D100 D103 N802
     medusa/sab.py D100 D202 D400 D401 N802 N806
     medusa/sbdatetime.py D100 D101 D102 D105 D202 D205 D400 N801
