+++ runscripts/init.fedora
@@ -48,7 +48,7 @@ fi
 start() {
         # Start daemon.
         echo -n $"Starting $prog: "
-        daemon --user=${username} --pidfile=${pidfile} ${nice} ${python_bin} ${homedir}/SickBeard.py ${options}
+        daemon --user=${username} --pidfile=${pidfile} ${nice} ${python_bin} ${homedir}/start.py ${options}
         RETVAL=$?
         echo
         [ $RETVAL -eq 0 ] && touch $lockfile
