+++ runscripts/init.systemd
@@ -13,7 +13,7 @@
 #    - Adjust ExecStart= to point to your python and Medusa executables.
 #       The FIRST token of the command line must be an ABSOLUTE FILE NAME,
 #       then followed by arguments for the process.
-#       If no --datadir is given, data is stored in same dir as SickBeard.py
+#       If no --datadir is given, data is stored in same dir as start.py
 #       Arguments can also be set in EnvironmentFile (except python)
 #
 #    - WantedBy= specifies which target (i.e. runlevel) to start Medusa for.
@@ -24,21 +24,21 @@
 ### Example Using Medusa as daemon with pid file
 # Type=forking
 # PIDFile=/var/run/PyMedusa/Medusa.pid
-# ExecStart=/usr/bin/python2.7 /opt/medusa/SickBeard.py -q --daemon --nolaunch --pidfile=/var/run/PyMedusa/Medusa.pid --datadir=/opt/medusa
+# ExecStart=/usr/bin/python2.7 /opt/medusa/start.py -q --daemon --nolaunch --pidfile=/var/run/PyMedusa/Medusa.pid --datadir=/opt/medusa
 
 ## Example Using Medusa as daemon without pid file
 # Type=forking
 # GuessMainPID=no
-# ExecStart=/usr/bin/python2.7 /opt/medusa/SickBeard.py -q --daemon --nolaunch --datadir=/opt/medusa
+# ExecStart=/usr/bin/python2.7 /opt/medusa/start.py -q --daemon --nolaunch --datadir=/opt/medusa
 
 ### Example Using simple
 # Type=simple
-# ExecStart=/usr/bin/python2.7 /opt/medusa/SickBeard.py -q --nolaunch
+# ExecStart=/usr/bin/python2.7 /opt/medusa/start.py -q --nolaunch
 
 ### Example Using simple with EnvironmentFile where APP_DATA=/home/medusa/.medusa in /etc/medusa.conf
 # Type=simple
 # EnvironmentFile=/etc/medusa.conf
-# ExecStart=/usr/bin/python2.7 /opt/medusa/SickBeard.py -q --nolaunch --datadir=${APP_DATA}
+# ExecStart=/usr/bin/python2.7 /opt/medusa/start.py -q --nolaunch --datadir=${APP_DATA}
 
 ### Configuration
 
@@ -51,7 +51,7 @@ Group=medusa
 
 Type=forking
 GuessMainPID=no
-ExecStart=/usr/bin/python2.7 /opt/medusa/SickBeard.py -q --daemon --nolaunch --datadir=/opt/medusa
+ExecStart=/usr/bin/python2.7 /opt/medusa/start.py -q --daemon --nolaunch --datadir=/opt/medusa
 
 [Install]
 WantedBy=multi-user.target
